Create richer notifications using available APIs.
On Jelly Bean and above, create a BigPictureStyle notification, which expands to
show the thumbnail of the current episode. On Honeycomb and above, use that
thumbnail as the largeIcon. Notifications on devices below Honeycomb are
unchanged.
